+++ gui/src/main/java/io/xj/gui/controllers/MainPaneTopController.java
@@ -5,18 +5,12 @@ package io.xj.gui.controllers;
 import io.xj.gui.services.FabricationService;
 import io.xj.gui.services.FabricationStatus;
 import io.xj.gui.services.LabService;
-import io.xj.gui.services.LabStatus;
-import io.xj.nexus.InputMode;
-import io.xj.nexus.OutputFileMode;
-import io.xj.nexus.OutputMode;
 import javafx.beans.binding.Bindings;
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.value.ObservableValue;
-import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.*;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.layout.VBox;
 import org.springframework.stereotype.Service;
 
@@ -25,7 +19,6 @@ import java.util.List;
 
 @Service
 public class MainPaneTopController extends VBox implements ReadyAfterBootController {
-  static final int FABRICATION_CONFIG_VIEW_HEIGHT = 240;
   static final List<FabricationStatus> BUTTON_ACTION_ACTIVE_IN_FABRICATION_STATES = Arrays.asList(
     FabricationStatus.Standby,
     FabricationStatus.Active,
@@ -37,9 +30,9 @@ public class MainPaneTopController extends VBox implements ReadyAfterBootControl
   final static String BUTTON_TEXT_STOP = "Stop";
   final static String BUTTON_TEXT_RESET = "Reset";
   final FabricationService fabricationService;
+  final ModalFabricationSettingsController modalFabricationSettingsController;
   final ModalLabAuthenticationController modalLabAuthenticationController;
   final LabService labService;
-  final BooleanProperty configVisible = new SimpleBooleanProperty(false);
 
   @FXML
   Button buttonAction;
@@ -48,69 +41,20 @@ public class MainPaneTopController extends VBox implements ReadyAfterBootControl
   Label labelFabricationStatus;
 
   @FXML
-  ToggleButton toggleShowConfig;
-
-  @FXML
-  VBox fabricationConfigView;
-
-  @FXML
-  TextField fieldInputTemplateKey;
-
-  @FXML
-  ChoiceBox<InputMode> choiceInputMode;
-
-  @FXML
-  ChoiceBox<OutputMode> choiceOutputMode;
-
-  @FXML
-  ChoiceBox<OutputFileMode> choiceOutputFileMode;
-
-  @FXML
-  Label labelInputMode;
-
-  @FXML
-  Label labelOutputFileMode;
-
-  @FXML
-  TextField fieldOutputSeconds;
-
-  @FXML
-  Label labelOutputSeconds;
-
-  @FXML
-  TextField fieldOutputPathPrefix;
-
-  @FXML
-  Label labelOutputPathPrefix;
-
-  @FXML
-  TextField fieldCraftAheadSeconds;
-
-  @FXML
-  TextField fieldDubAheadSeconds;
-
-  @FXML
-  TextField fieldShipAheadSeconds;
-
-  @FXML
-  TextField fieldOutputChannels;
-
-  @FXML
-  TextField fieldOutputFrameRate;
-
-  @FXML
   public Button buttonLab;
 
   @FXML
   public Label labelLabStatus;
 
   public MainPaneTopController(
-    FabricationService fabricationService,
+    ModalFabricationSettingsController modalFabricationSettingsController,
     ModalLabAuthenticationController modalLabAuthenticationController,
+    FabricationService fabricationService,
     LabService labService
   ) {
-    this.fabricationService = fabricationService;
+    this.modalFabricationSettingsController = modalFabricationSettingsController;
     this.modalLabAuthenticationController = modalLabAuthenticationController;
+    this.fabricationService = fabricationService;
     this.labService = labService;
   }
 
@@ -120,43 +64,9 @@ public class MainPaneTopController extends VBox implements ReadyAfterBootControl
     buttonAction.textProperty().bind(Bindings.createStringBinding(this::computeActionButtonText, fabricationService.statusProperty()));
     fabricationService.statusProperty().addListener(this::handleFabricationStatusChange);
 
-    // Input mode is locked in PRODUCTION unless we are connected to a Lab
-    choiceInputMode.valueProperty().bindBidirectional(fabricationService.inputModeProperty());
-    choiceInputMode.disableProperty().bind(labService.statusProperty().isEqualTo(LabStatus.Authenticated).not());
-    labelInputMode.disableProperty().bind(labService.statusProperty().isEqualTo(LabStatus.Authenticated).not());
-    choiceInputMode.setItems(FXCollections.observableArrayList(InputMode.PRODUCTION));
-    labService.statusProperty().addListener(this::handleLabStatusChange);
-    fieldInputTemplateKey.textProperty().bindBidirectional(fabricationService.inputTemplateKeyProperty());
-
-    choiceOutputMode.getItems().setAll(OutputMode.values());
-    choiceOutputMode.valueProperty().bindBidirectional(fabricationService.outputModeProperty());
-
-    choiceOutputFileMode.getItems().setAll(OutputFileMode.values());
-    choiceOutputFileMode.valueProperty().bindBidirectional(fabricationService.outputFileModeProperty());
-    choiceOutputFileMode.disableProperty().bind(fabricationService.outputModeProperty().isEqualTo(OutputMode.FILE).not());
-    labelOutputFileMode.disableProperty().bind(fabricationService.outputModeProperty().isEqualTo(OutputMode.FILE).not());
-
-    fieldOutputSeconds.textProperty().bindBidirectional(fabricationService.outputSecondsProperty());
-    fieldOutputSeconds.disableProperty().bind(fabricationService.outputModeProperty().isEqualTo(OutputMode.FILE).not());
-    labelOutputSeconds.disableProperty().bind(fabricationService.outputModeProperty().isEqualTo(OutputMode.FILE).not());
-
-    fieldOutputPathPrefix.textProperty().bindBidirectional(fabricationService.outputPathPrefixProperty());
-    fieldOutputPathPrefix.disableProperty().bind(fabricationService.outputModeProperty().isEqualTo(OutputMode.FILE).not());
-    labelOutputPathPrefix.disableProperty().bind(fabricationService.outputModeProperty().isEqualTo(OutputMode.FILE).not());
-
-    fieldCraftAheadSeconds.textProperty().bindBidirectional(fabricationService.craftAheadSecondsProperty());
-    fieldDubAheadSeconds.textProperty().bindBidirectional(fabricationService.dubAheadSecondsProperty());
-    fieldShipAheadSeconds.textProperty().bindBidirectional(fabricationService.shipAheadSecondsProperty());
-    fieldOutputFrameRate.textProperty().bindBidirectional(fabricationService.outputFrameRateProperty());
-    fieldOutputChannels.textProperty().bindBidirectional(fabricationService.outputChannelsProperty());
-
     labelFabricationStatus.textProperty().bind(fabricationService.statusProperty().map(Enum::toString).map((status) -> String.format("Fabrication %s", status)));
 
-    toggleShowConfig.setSelected(configVisible.get());
-
     labelLabStatus.textProperty().bind(labService.statusProperty().map(Enum::toString));
-
-    updateConfigVisibility();
   }
 
   @Override
@@ -186,9 +96,8 @@ public class MainPaneTopController extends VBox implements ReadyAfterBootControl
   }
 
   @FXML
-  public void toggleShowConfig(ActionEvent ignored) {
-    configVisible.set(toggleShowConfig.isSelected());
-    updateConfigVisibility();
+  public void handleShowFabricationSettings(ActionEvent ignored) {
+    modalFabricationSettingsController.launchModal();
   }
 
   @FXML
@@ -197,20 +106,6 @@ public class MainPaneTopController extends VBox implements ReadyAfterBootControl
       labService.launchInBrowser();
     } else {
       modalLabAuthenticationController.launchModal();
-    }
-  }
-
-  void updateConfigVisibility() {
-    if (configVisible.get()) {
-      fabricationConfigView.setVisible(true);
-      fabricationConfigView.setMinHeight(FABRICATION_CONFIG_VIEW_HEIGHT);
-      fabricationConfigView.setPrefHeight(FABRICATION_CONFIG_VIEW_HEIGHT);
-      fabricationConfigView.setMaxHeight(FABRICATION_CONFIG_VIEW_HEIGHT);
-    } else {
-      fabricationConfigView.setVisible(false);
-      fabricationConfigView.setMinHeight(0);
-      fabricationConfigView.setPrefHeight(0);
-      fabricationConfigView.setMaxHeight(0);
     }
   }
 
@@ -231,18 +126,5 @@ public class MainPaneTopController extends VBox implements ReadyAfterBootControl
 
   private Boolean isActionButtonActive() {
     return BUTTON_ACTION_ACTIVE_IN_FABRICATION_STATES.contains(fabricationService.statusProperty().get());
-  }
-
-  private void handleLabStatusChange(ObservableValue<? extends LabStatus> observable, LabStatus oldValue, LabStatus newValue) {
-    if (newValue == LabStatus.Authenticated) {
-      if (choiceInputMode.getItems().size() == 1) {
-        choiceInputMode.getItems().add(InputMode.PREVIEW);
-      }
-    } else {
-      if (choiceInputMode.getItems().size() == 2) {
-        choiceInputMode.getItems().remove(1);
-      }
-      choiceInputMode.setValue(InputMode.PRODUCTION);
-    }
   }
 }
