+++ gui/src/main/resources/styles/default-theme.css
@@ -41,10 +41,6 @@
   -fx-background-color: #c9c9c9;
 }
 
-#fabricationConfigView {
-  -fx-background-color: #b0b0b0;
-}
-
 .main-timeline,
 .timeline-position-spacer {
   -fx-background-color: #e0e0e0;
