+++ gui/src/main/resources/views/modal-fabrication-settings.fxml
@@ -1,0 +1,149 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!-- Copyright (c) XJ Music Inc. (https://xjmusic.com) All Rights Reserved. -->
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.Priority?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox spacing="10" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="io.xj.gui.controllers.ModalFabricationSettingsController">
+  <padding>
+    <Insets bottom="15" left="15" right="15" top="15" /> <!-- Padding values -->
+  </padding>
+
+  <VBox BorderPane.alignment="CENTER">
+    <VBox.vgrow>
+      <Priority fx:constant="ALWAYS" />
+    </VBox.vgrow>
+
+    <HBox alignment="CENTER_LEFT" spacing="10">
+      <padding>
+        <Insets left="10" right="10" top="10" />
+      </padding>
+      <VBox alignment="CENTER_LEFT" prefHeight="60.0">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Input Template ID" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldInputTemplateKey" prefWidth="300" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT" prefHeight="60.0">
+        <Label fx:id="labelInputMode" alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Input Mode" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <ChoiceBox fx:id="choiceInputMode" prefHeight="25.0" prefWidth="120" />
+      </VBox>
+    </HBox>
+
+    <HBox alignment="CENTER_LEFT" spacing="10">
+      <padding>
+        <Insets left="10" right="10" top="10" />
+      </padding>
+      <VBox alignment="CENTER_LEFT" prefHeight="60.0">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Output Mode" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <ChoiceBox fx:id="choiceOutputMode" prefHeight="25.0" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT" prefHeight="41.0">
+        <Label fx:id="labelOutputFileMode" alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Output File Mode" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <ChoiceBox fx:id="choiceOutputFileMode" prefHeight="25.0" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT" prefHeight="60">
+        <Label fx:id="labelOutputSeconds" alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Output File Seconds" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldOutputSeconds" prefHeight="25.0" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT" prefHeight="41.0">
+        <Label fx:id="labelOutputPathPrefix" alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Output File Path Prefix" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldOutputPathPrefix" prefHeight="25.0" />
+      </VBox>
+    </HBox>
+
+    <HBox alignment="CENTER_LEFT" prefHeight="41.0" spacing="10">
+      <padding>
+        <Insets left="10" right="10" top="10" />
+      </padding>
+      <VBox alignment="CENTER_LEFT">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Output Frame Rate" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldOutputFrameRate" prefHeight="25.0" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Output Channels" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldOutputChannels" prefHeight="25.0" />
+      </VBox>
+      <Pane prefWidth="20" />
+      <VBox alignment="CENTER_LEFT">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Craft Ahead Seconds" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldCraftAheadSeconds" prefHeight="25.0" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Dub Ahead Seconds" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldDubAheadSeconds" prefHeight="25.0" />
+      </VBox>
+      <VBox alignment="CENTER_LEFT">
+        <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" text="Ship Ahead Seconds" textAlignment="RIGHT">
+          <graphic>
+            <Pane prefWidth="15.0" />
+          </graphic>
+        </Label>
+        <Pane prefWidth="15.0" />
+        <TextField fx:id="fieldShipAheadSeconds" prefHeight="25.0" />
+      </VBox>
+    </HBox>
+  </VBox>
+
+  <HBox alignment="BOTTOM_LEFT" prefHeight="50.0" prefWidth="640.0">
+    <Pane layoutX="192.0" layoutY="10.0" prefHeight="50.0" prefWidth="9.0" />
+    <Button fx:id="buttonClose" onAction="#handleClose" prefHeight="25.0" prefWidth="56.0" text="_OK" />
+  </HBox>
+</VBox>
